# api.yaml
openapi: 3.1.1
info:
  title: Doctree API
  description: API для иерархической системы хранения документов
  version: "2.0.0"
servers:
  - url: /api/v2
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GetFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        fileUrl:
          type: string
    DocumentRelation:
      type: object
      properties:
        documentId:
          type: string
          format: uuid
        type:
          type: string
    DocumentNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/GetFile"
        relations:
          type: array
          items:
            $ref: "#/components/schemas/DocumentRelation"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/DocumentNode"
      required: [id, title]
    DocumentWithNodes:
      type: object
      properties:
        nodeTitle:
          type: string
          format: uuid
        documents:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              title:
                type: string
              description:
                type: string
              tags:
                type: array
                items:
                  type: string
              files:
                type: array
                items:
                  $ref: "#/components/schemas/GetFile"
      required: [id, title]
...
    CreateDocumentRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title]
    UpdateDocumentRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
...
security:
  - bearerAuth: []
paths:
...
  /docs:
    post:
      tags: [Document]
      summary: Создать документ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Документ создан
        "403":
            description: У пользователя нет прав
        "400":
          description: Некорректный запрос
        
    get:
      tags: [Document]
      summary: Поиск документов
      parameters:
        - in: query
          name: title
          schema:
            type: string
        - in: query
          name: description
          schema:
            type: string
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: text
          schema:
            type: string
      responses:
        "200":
          description: Найденные документы
          content:
            application/json:
              schema:
                type: object
                properties:
                  docs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
  /docs/{id}:
    get:
      tags: [Document]
      summary: Получить документ по id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Документ найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  doc:
                    $ref: "#/components/schemas/Document"
        "400":
          description: Некорректный запрос
        
        "403":
            description: У пользователя нет прав
        
        "404":
          description: Документ не найден
        
    patch:
      tags: [Document]
      summary: Обновить документ
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentRequest"
      responses:
        "200":
          description: Документ обновлен
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Документ не найден
    delete:
      tags: [Document]
      summary: Удалить документ
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Документ удален
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Документ не найден
  /docs/nodes/{nodeId}:
    get:
      tags: [Document]
      summary: Получить все документы, привязанные к узлу
      parameters:
        - in: path
          name: nodeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Найденные документы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentWithNodes"
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Узел не найден
  /docs/{id}/files:
    patch:
      tags: [Document]
      summary: Загрузить файл к документу
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "201":
          description: Файл прикреплен
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Документ не найден
  /docs/{id}/files/{fileId}:
    delete:
      tags: [Document]
      summary: Отвязать файл от документа
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: fileId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Файл отвязан
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Документ или файл не найдены
  /docs/{id}/nodes/{nodeId}:
    patch:
      tags: [Document]
      summary: Привязать документ к узлу
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: nodeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Документ отвязан от другого узла иерархии и привязан к заданному
        "201":
          description: Документ привязан к узлу
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Документ или узел не найден
    delete:
      tags: [Document]
      summary: Отвязать документ от узла
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: nodeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Документ отвязан от узла
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Документ или узел не найден
  /docs/{id}/relations/{relatedId}:
    patch:
      tags: [Document]
      summary: Связать документ с другим документом
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: relatedId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Связь создана
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Документ или узлы не найдены
    delete:
      tags: [Document]
      summary: Удалить связь между документами
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: relatedId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Связь удалена
        "400":
          description: Некорректный запрос
        "403":
            description: У пользователя нет прав
        "404":
          description: Документ или связь не найдены
...